refactor(mespapiers): Move all Scan's components in specific folder
For better readability and maintainability
