+++ packages/cozy-mespapiers-lib/src/components/ModelSteps/Scan/ScanActions/ScanFlagshipActions.jsx
